Added formatting for miles + 1 issue
Station -> Map, makes Station show double.

+++ Pacers_BikeShare/app/src/main/java/com/misabelleeli/pacers_bikeshare/StationFragment.java
@@ -185,7 +185,7 @@ public class StationFragment extends Fragment implements CompoundButton.OnChecke
         super.onViewCreated(view, savedInstanceState);
     }
 
-    public static void populateStations(String stationName, String addr, int bike, int dock, int miles) {
+    public static void populateStations(String stationName, String addr, int bike, int dock, float miles) {
 
         stations.add(new Station(stationName, addr, bike, dock, miles));
